Fix performance when reading job LastRun
The problem is that the code was repetitively trying to get subscription ID for all scheduled jobs. It means that for many jobs it is called really many times. I have rewritten the SQL so it returns the required information using JOIN, which is much faster.

Also I have implemented a memory cache per subscription ID, because if we have n jobs for every subscription it doesn't make sense to fetch everything n times. The audit log can be large especially if we are talking about daily jobs.