Add file name and executable file validation.
Added a check to ensure that the file name does not contain more than one period. If it does, an error message is set in `TempData` and the user is redirected to the "Index" action.

Introduced a new block of code that reads the first two bytes of the uploaded file to determine if it is an executable file (by checking for the "MZ" header). If the file is identified as an executable, an error message is set in `TempData` and the user is redirected to the "Index" action.